# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Prayag Polymers pricing-intelligence API (product catalog, MRP history, competition analysis)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /price-finder/search:
    get:
      operationId: getPriceFinderSearch
      tags: [catalog]
      summary: Progressively filter Price Finder products by spoken or typed terms
      parameters:
        - name: q
          in: query
          required: false
          description: Backwards-compatible single filter term.
          schema:
            type: string
        - name: filters
          in: query
          required: false
          description: Accumulating filter phrases. Each must match a product field.
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 3000
            default: 3000
      responses:
        "200":
          description: Matching products, live count, and next-filter suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceFinderSearchResponse"
  /price-finder/browse:
    get:
      operationId: getPriceFinderBrowse
      tags: [catalog]
      summary: Browse the catalogue by division and category
      parameters:
        - name: division
          in: query
          required: false
          schema:
            type: string
        - name: category
          in: query
          required: false
          description: Use "__uncategorised__" for blank categories.
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 300
            default: 200
      responses:
        "200":
          description: Browse options and, when selected, products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceFinderBrowseResponse"
  /price-finder/product/{itemCode}:
    get:
      operationId: getPriceFinderProduct
      tags: [catalog]
      summary: Get current Prayag and competitor prices for one product
      parameters:
        - name: itemCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Price Finder product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceFinderProductResponse"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    PriceFinderSearchResult:
      type: object
      required:
        - itemCode
        - productName
        - division
        - category
        - currentMrp
        - currentEffectiveDate
        - currentValidTo
        - upcomingMrp
        - upcomingEffectiveDate
        - upcomingChangePct
        - hasCompetitorData
        - discontinuedFrom
      properties:
        itemCode:
          type: string
        productName:
          type: ["string", "null"]
        division:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        currentMrp:
          type: ["number", "null"]
        currentEffectiveDate:
          type: ["string", "null"]
          description: First day the current MRP is valid (YYYY-MM-DD).
        currentValidTo:
          type: ["string", "null"]
          description: Final day the current MRP is valid, or null when open-ended.
        upcomingMrp:
          type: ["number", "null"]
          description: Earliest scheduled MRP after today, or null.
        upcomingEffectiveDate:
          type: ["string", "null"]
        upcomingChangePct:
          type: ["number", "null"]
          description: One-decimal change from current to upcoming MRP; zero is retained.
        hasCompetitorData:
          type: boolean
        discontinuedFrom:
          type: ["string", "null"]
          format: date
    PriceFinderSearchResponse:
      type: object
      required: [results, totalCount, filterSuggestions, unmatchedFilters]
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/PriceFinderSearchResult"
        totalCount:
          type: integer
        filterSuggestions:
          type: array
          items:
            $ref: "#/components/schemas/PriceFinderFilterSuggestion"
        unmatchedFilters:
          type: array
          description: Terms rejected during progressive narrowing. Results and totalCount retain the last successful filter set.
          items:
            type: string
    PriceFinderFilterSuggestion:
      type: object
      required: [field, distinctCount, values]
      properties:
        field:
          type: string
        distinctCount:
          type: integer
        values:
          type: array
          items:
            type: object
            required: [value, count]
            properties:
              value:
                type: string
              count:
                type: integer
    PriceFinderBrowseDivision:
      type: object
      required: [division, count]
      properties:
        division:
          type: string
        count:
          type: integer
    PriceFinderBrowseCategory:
      type: object
      required: [category, label, count]
      properties:
        category:
          type: ["string", "null"]
        label:
          type: string
        count:
          type: integer
    PriceFinderBrowseResponse:
      type: object
      required: [divisions, categories, products]
      properties:
        divisions:
          type: array
          items:
            $ref: "#/components/schemas/PriceFinderBrowseDivision"
        categories:
          type: array
          items:
            $ref: "#/components/schemas/PriceFinderBrowseCategory"
        products:
          type: array
          items:
            $ref: "#/components/schemas/PriceFinderSearchResult"
    PriceFinderProduct:
      type: object
      required:
        - itemCode
        - productName
        - division
        - category
        - currentMrp
        - currentEffectiveDate
        - currentValidTo
        - upcomingMrp
        - upcomingEffectiveDate
        - upcomingChangePct
        - discontinuedFrom
      properties:
        itemCode:
          type: string
        productName:
          type: ["string", "null"]
        division:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        currentMrp:
          type: ["number", "null"]
        currentEffectiveDate:
          type: ["string", "null"]
          description: First day the current MRP is valid (YYYY-MM-DD).
        currentValidTo:
          type: ["string", "null"]
          description: Final day the current MRP is valid, or null when open-ended.
        upcomingMrp:
          type: ["number", "null"]
        upcomingEffectiveDate:
          type: ["string", "null"]
        upcomingChangePct:
          type: ["number", "null"]
          description: One-decimal change from current to upcoming MRP; zero is retained.
        discontinuedFrom:
          type: ["string", "null"]
          format: date
    PriceFinderCompetitor:
      type: object
      required:
        - competitor
        - price
        - effectiveDate
        - validTo
        - priceBasis
        - upcomingPrice
        - upcomingEffectiveDate
        - upcomingChangePct
        - gapPct
        - message
      properties:
        competitor:
          type: string
        price:
          type: number
        effectiveDate:
          type: ["string", "null"]
          description: First day the current competitor price is valid (YYYY-MM-DD).
        validTo:
          type: ["string", "null"]
          description: Final day the current competitor price is valid, or null when open-ended.
        priceBasis:
          type: ["string", "null"]
        upcomingPrice:
          type: ["number", "null"]
          description: Earliest scheduled competitor price after today, on the same comparison basis.
        upcomingEffectiveDate:
          type: ["string", "null"]
        upcomingChangePct:
          type: ["number", "null"]
          description: One-decimal change from current to upcoming price; zero is retained.
        gapPct:
          type: ["number", "null"]
        message:
          type: ["string", "null"]
    PriceFinderProductResponse:
      type: object
      required: [product, competitors]
      properties:
        product:
          $ref: "#/components/schemas/PriceFinderProduct"
        competitors:
          type: array
          items:
            $ref: "#/components/schemas/PriceFinderCompetitor"
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key